requirements.txt
torch
torchvision
numpy
opencv-python
pillow
matplotlib
nltk
scikit-image
tqdm
